+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Fixed bug in `sortx` in `pirateplot()`. Sorting data by functions (e.g. `sortx = "mean"`) should now work.
 
+* Added `gl` argument to `pirateplot()` to specify locations of gridlines
+
 # yarrr 0.1.4
 
 * Some minor changes to default plotting parameters that I think make the default plots look a bit nicer.